vim: uniform tab with in tabbar
flipping through tabs has a weird behavior of looking odd because tabs
change width: when it gets to the last tab, it may become longer than
the line. even without it, the tabs "dance" as one flips through them
due to that change of length.

this is fixed by making all tabs end with either space, '*' for visible
buffers (current behavior), '+' for modified buffers (current behavior)
and '&' for visible and modified buffers (instead of the current '*+').

Signed-off-by: Idan Kedar <idank@primarydata.com>